mergetwosortedlists/main.go
package mergetwosortedlists

/*
https://leetcode.com/problems/merge-two-sorted-lists/

Merge two sorted linked lists and return it as a new list. The new list should be made by splicing together the nodes of the first two lists.

Example:

Input: 1->2->4, 1->3->4
Output: 1->1->2->3->4->4

Approach:

* While current nodes are non-nil
* Picks whichever node is smaller and adds that to list
* Increment node that was chosen and merged list node
* If a node is non-nil, that is the node that the rest of the list should reference
*/

// ListNode is a node in a singly-linked list
type ListNode struct {
	Val  int
	Next *ListNode
}

func mergeTwoLists(l1 *ListNode, l2 *ListNode) *ListNode {
	merged := ListNode{}
	current := &merged

	for l1 != nil && l2 != nil {
		if l1.Val <= l2.Val {
			current.Next = l1
			l1 = l1.Next
		} else {
			current.Next = l2
			l2 = l2.Next
		}
		current = current.Next
	}

	if l1 != nil {
		current.Next = l1
	}

	if l2 != nil {
		current.Next = l2
	}

	return merged.Next
}
